+++ NEWS.md
@@ -3,6 +3,10 @@
 * `vl_generators()` now gives correct results for sets which fail `optc_test()`
 * `minimize_vl()` now returns better results when `method="hamming"` by allowing for voice crossings (#4).
 
+## More flexible handling of OPTIC symmetries
+* Set theory functions including `tn()`, `startzero()`, and so on gain an `optic` parameter which allows the user
+  to specify which symmetries to consider.
+
 # musicMCT 0.2.0
 
 ## New ineqmat features